+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Raised minimum required PHP version to 7.3
 * Added more type hints
 * `stubbles\webapp\response\mimetypes\Csv` will now throw an exception when a line can't be serialized instead of silently converting to an empty line
+* `stubbles\webapp\htmlpassthrough\HtmlFilePassThrough` now serves an error 500 in case the file can't be read
 
 ## 7.0.0 (2016-08-06)